refactored app.ts and routes/index.ts
