+++ server/package.json
@@ -36,6 +36,7 @@
     "yarn": "^1.22.4"
   },
   "devDependencies": {
+    "@types/cors": "^2.8.12",
     "@types/express": "^4.17.13",
     "@types/jest": "^26.0.24",
     "@types/supertest": "^2.0.11",
