--- server/src/routes/index.js
@@ -1,19 +1,0 @@
-const fs = require('fs')
-const path = require('path')
-
-const routes = {}
-const basename = path.basename(__filename)
-
-fs.readdirSync(__dirname)
-  .filter((file) => {
-    return (
-      file.indexOf('.') !== 0 && file !== basename && file.slice(-3) === '.js'
-    )
-  })
-  .forEach((file) => {
-    const route = require(path.join(__dirname, file))
-    const name = file.split('.')[0]
-    routes[name] = route
-  })
-
-module.exports = routes
